make compatible with g++ 12 and clang++ 15

+++ src/Data2D.cpp
@@ -42,7 +42,8 @@ std::vector<std::pair<double, double>> sciformats::jdx::Data2D::parseXppYYData(
     uint64_t count = 0;
     for (auto yRaw : yData)
     {
-        auto x = firstX + nominator / denominator * count++;
+        auto x = firstX + nominator /
+            static_cast<double>(denominator) * static_cast<double>(count++);
         auto y = yFactor * yRaw;
         xyData.emplace_back(x, y);
     }
